Change helper layer provided by the helper class.
